requirements.txt
aniso8601==1.1.0
backports-abc==0.4
beautifulsoup4==4.5.0
bs4==0.0.1
click==6.6
configparser==3.5.0
coverage==4.2
et-xmlfile==1.0.1
Flask==0.11.1
Flask-RESTful==0.3.5
Flasky==0.1.0
itsdangerous==0.24
jdcal==1.2
Jinja2==2.8
MarkupSafe==0.23
nose==1.3.7
numpy==1.11.1
openpyxl==2.3.5
pandas==0.18.1
Pillow==3.3.0
PyMySQL==0.7.6
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.10.0
six==1.10.0
SQLAlchemy==1.0.15
tornado==4.4.1
Werkzeug==0.11.10
